requirements.txt
flask
nosetest